requirements.txt
hydra-core
paho-mqtt
pandas